added module api.js
