+++ js/data.js
@@ -1,4 +1,4 @@
-import {QUANTITY_PHOTOS, MIN_LIKES, MAX_LIKES, MIN_COMMENTS, MAX_COMMENTS, QUANTITY_AVATAR, NAMES, DESCRIPTION, MESSAGES} from './constants.js';
+// import {QUANTITY_PHOTOS, MIN_LIKES, MAX_LIKES, MIN_COMMENTS, MAX_COMMENTS, QUANTITY_AVATAR, NAMES, DESCRIPTION, MESSAGES} from './constants.js';
 
 /*import {getRandomInteger, generateCountOfPhotos, getRandomArrayElement,createUniqueId} from './utils.js';
